% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iupMLE.R
\name{iupMLE}
\alias{iupMLE}
\title{Maximum likelihood estimator for infectious units per million (IUPM) to incorporate partial deep sequencing information}
\usage{
iupMLE(data, maxit = 10000, lb = 1e-06, ub = Inf)
}
\arguments{
\item{data}{Assay data, with rows representing the distinct viral lineages (DVL) and columns representing the wells. Any columns with \code{NA} values will be treated as positive but missing deep sequencing information.}

\item{maxit}{The maximum number of iterations (passed to \code{optim}). Default is \code{maxit = 1E4}.}

\item{lb}{Lower-bound on the IUPM (passed to \code{optim}). Default is \code{lb = 1E-6}.}

\item{ub}{Upper-bound on the IUPM (passed to \code{optim}). Default is \code{ub = Inf}.}
}
\value{
A named list with the following slots:
\item{mle}{MLE of the IUPM}
\item{bc_mle}{Bias-corrected MLE of the IUPM}
\item{se}{Standard error for the MLE of the IUPM}
\item{se_bc}{Standard error for the bias-corrected MLE of the IUPM}
}
\description{
Maximum likelihood estimator for infectious units per million (IUPM) to incorporate partial deep sequencing information
}
