% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_assay.R
\name{simulate_assay}
\alias{simulate_assay}
\title{Simulate a dilution assay with partial deep sequencing information}
\usage{
simulate_assay(M, n, lambda, q, remove_undetected = TRUE)
}
\arguments{
\item{M}{Total number of wells.}

\item{n}{Total number of underlying distinct viral lineages (DVL) to detect}

\item{lambda}{A vector of length \code{n} with the expected number of infected cells per DVL}

\item{q}{Proportion of p24-positive wells to be deep sequenced}

\item{remove_undetected}{(Logical) If \code{remove_undetected = TRUE} (the default), DVL that were not detected in any wells are removed}
}
\value{
A matrix with \code{n} rows (one per DVL) and \code{M} columns (one per well)
}
\description{
Simulate a dilution assay with partial deep sequencing information
}
